Merge branch 'mv88e6390-missing-phy-ID'
Andrew Lunn says:

====================
Work around missing PHY prodcut ID in mv88e6390

The internal PHYs of the MV88E6390 have a Marvell OUI, but the product
ID is zero. Work around this by trapping reads to the ID, and if it is
zero, return the MV88E6390 family ID.

v2: Move the workaround into the central mdio read function.
    Enable the temperature sensor, even though it does not work on the 6390,
    but it does on the 6341, which has the same ID problem.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 2a7b6216692cd66f4c94e6ebb48e43038349ca7d ]